refactor(data): 迁移 CR3BP 平动点枚举 (#438)
