+++ e2m2e/data/templates/__init__.py
@@ -17,6 +17,7 @@ from .enums import (
     BoundaryMode,
     ConvergenceState,
     FailureCause,
+    LibrationPoint,
     OrbitFamilyType,
     ProjectionPlane,
     ReferenceFrame,
@@ -47,6 +48,7 @@ __all__ = [
     "BoundaryMode",
     "ConvergenceState",
     "FailureCause",
+    "LibrationPoint",
     "OrbitFamilyType",
     "StabilityLabel",
     "BifurcationLabel",
