+++ tests/algorithm/normal_form/test_qf_multipoint.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-from e2m2e.algorithm.dynamics import CR3BP_System, LibrationPoint
+from e2m2e.algorithm.dynamics import CR3BP_System
 from e2m2e.algorithm.normal_form import NormalFormContext
 from e2m2e.algorithm.normal_form.quasi_floquet import (
     J6,
@@ -19,6 +19,7 @@ from e2m2e.algorithm.normal_form.quasi_floquet import (
     _solve_qf_multipoint,
     real_normal_form_matrix,
 )
+from e2m2e.data.templates.enums import LibrationPoint
 
 pytestmark = pytest.mark.theory
